Logging improvements (#229)
* Fix for newer eclipse (still works when compiling with Mars)

* Don't set the message out put level

For some reason, this interferes with the logging console.

* Don't filter out a log message

* Update Launcher UI to support selecting logging implementation

* Add all the logging implementations from MyBatis

Also setup the same order of checking for implementations

* Cleanup unused imports

* Eclipse doesn't bundle log4j2, so remove it

It's too much of a hassle to support log4j2 in Eclipse

* Update Button options, add forceXXXlogging methods

* Hook up the logging UI with the Ant task

* Add Log4j2 back in - had a eureka moment with the eclipse build

* Need to catch throwable instead of exception

If the logger implementation isn't found, there will be a
NoClassDefFoundError

* Change the eclipse build for the core library

This change means makes it so that all the logger implementations don't
end up on the plugin classpath.  If a specific logger is desired, it
needs to be added to the launch configuration classpath.

* Minor cleanup

* Fix logging in the Eclipse feature

Add the logging classes to the Ant support library so they will be
loaded by the Ant classloader rather than the plugin classloader.  THis
way they will pickup build path changes from the launch configuration.

* Add a helpful message if the loggin implementation is missing

* NLS