[jdk] Correct java method version usage - supports java 6
System.lineSeparator is for java 7.  The code base is set for java 6.
Using java 6 will fail the build so this has been restored.  It appears
after testing that animal sniffer skips test side which mostly makes
sense.  If this same code was introduced on source side, animal sniffer
immediately fails the build.