[site] Findbugs is deprecated now, we switched to spotbugs but need old runs
- java 6 will run 2.5.5 still
- java 7 will run 3.0.5 (latest)
- java 8 will already be configured but needs split out for ibatis2
java2 module
- java 9 will already be configured

Note: The more broadly configured items for findbugs is now gone.  If we
want that we would need to add that back here.  Findbugs wile still
active has been a combination of others and I working to get core
changes back to that plugin as it had many bugs.  At this point
officially, spotbugs is the full replacement but does require java 8 to
run thus the profiles here.