Java and XML DOM Improvements and External DOM Renderers
Improve the Java and XML DOM classes. The Java DOM is made more
consistent and has additional capabilities so that most types of Java
code are now supported including unlimited nesting of classes, enums,
and interfaces.

The XML DOM now supports both SYSTEM and PUBLIC doc types.

For both, the rendering of the DOM classes to a persistent format has
been delegated to new renderer classes that are more accurate and
consistent than the prior method.

There are a few backwards incompatible changes that are documented in
the "What's New" page. These should not cause any difficulty for most
users.

Finally, added Kotlin support for test classes!